Rename AsyncGraphNode to GraphNodeResource
