+++ packages/react-graph-state/src/hooks/useGraphNodeSetValue.tsx
@@ -26,13 +26,13 @@
  * @copyright Alexis Munsayac 2020
  */
 import { useCallback } from 'react';
-import { useGraphLogic } from '../GraphLogic';
+import { useGraphStateDomain } from '../GraphStateDomain';
 import { GraphNode, GraphNodeDraftState } from '../graph-node';
 
 export default function useGraphNodeSetValue<T>(
   node: GraphNode<T>,
 ): (action: GraphNodeDraftState<T>) => void {
-  const logic = useGraphLogic();
+  const logic = useGraphStateDomain();
 
   return useCallback((action: GraphNodeDraftState<T>) => {
     logic.updateState(node, action);
